updated Sign-Up

+++ src/main/java/org/example/e_commerce/Service/CartService.java
@@ -47,7 +47,7 @@ public class CartService {
 
         if (!userOptional.isPresent()) {
             responseDTO.setMessage("User not found with username: " + username);
-            responseDTO.setStatusCode(-1L);
+            responseDTO.setStatusCode(-1);
             return responseDTO;
         }
 
@@ -60,7 +60,7 @@ public class CartService {
             cart.setUserid(userId);
             cartRepo.save(cart);
             responseDTO.setMessage("CART CREATED");
-            responseDTO.setStatusCode(0L);
+            responseDTO.setStatusCode(0);
             return responseDTO;
         }
 
@@ -69,7 +69,7 @@ public class CartService {
 
         if (product.getStockQuantity() < productRequestDTO.getQuantity()) {
             responseDTO.setMessage("Insufficient stock for product");
-            responseDTO.setStatusCode(-1L);
+            responseDTO.setStatusCode(-1);
             return responseDTO;
         }
 
@@ -81,7 +81,7 @@ public class CartService {
         cartDetailsRepo.save(cartDetails);
 
         responseDTO.setMessage("Products Added To Cart");
-        responseDTO.setStatusCode(0L);
+        responseDTO.setStatusCode(0);
         return responseDTO;
     }
 
@@ -100,7 +100,7 @@ public class CartService {
                 Product product = optionalProduct.get();
 
                 if (product.getStockQuantity() < productRequest.getQuantity()) {
-                    return new SignUpResponseDTO("Insufficient stock for product: " + productRequest.getProductId(), (long) HttpStatus.BAD_REQUEST.value());
+                    return new SignUpResponseDTO("Insufficient stock for product: " + productRequest.getProductId(), HttpStatus.BAD_REQUEST.value());
                 }
 
                 double amount = product.getPrice() * productRequest.getQuantity();
@@ -129,9 +129,9 @@ public class CartService {
             transaction.setAmount(totalAmount);
             transactionRepo.save(transaction);
 
-            return new SignUpResponseDTO("Purchase successful. Invoice Number: " + invoiceNumber, (long) HttpStatus.OK.value());
+            return new SignUpResponseDTO("Purchase successful. Invoice Number: " + invoiceNumber, HttpStatus.OK.value());
         } catch (RuntimeException e) {
-            return new SignUpResponseDTO("Error: " + e.getMessage(), (long) HttpStatus.INTERNAL_SERVER_ERROR.value());
+            return new SignUpResponseDTO("Error: " + e.getMessage(), HttpStatus.INTERNAL_SERVER_ERROR.value());
         }
     }
 }
